fix tests auth
