+++ src/user/schemas.py
@@ -109,19 +109,3 @@ class UsersProfileResponseSchema(UserProfileResponseSchema):
     model_config = ConfigDict(
         from_attributes=True,
     )
-
-
-# TESTS
-
-class UserResponseRefreshToken(UserSchema):
-    refresh_token: str
-
-
-class UserRequestEmailSchema(BaseModel):
-    email: EmailStr
-
-
-class UserRequestPasswordResetSchema(BaseModel):
-    reset_token: str
-    new_password: str
-    confirm_password: str
